api/requirements.txt
fastapi
uvicorn
requests
python-dotenv
rich
InquirerPy
httpx
